requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.100.2 
# openai==1.58.1
# LangGraph for interview flow
langgraph==0.6.3
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.6.3
langgraph-sdk==0.2.0
langchain==0.3.26
langchain-anthropic==0.3.18
langchain-chroma==0.2.5
langchain-core==0.3.74
langchain-deepseek==0.1.4
langchain-google-genai==2.1.9
langchain-openai==0.3.30
langchain-text-splitters==0.3.8
python-dotenv==1.0.0
python-multipart==0.0.6
jinja2==3.1.2
pymongo==4.6.1
motor==3.3.2
aiohttp==3.9.1
PyJWT==2.8.0
requests==2.31.0
asyncpg==0.29.0
# torch==2.1.2+cpu
# -f https://download.pytorch.org/whl/torch_stable.html
# sentence-transformers==2.2.2
psycopg2
# numpy==1.26.4
bcrypt==4.0.1
# Document processing
pypdf==3.17.4
python-docx==1.1.0
python-pptx==0.6.23
pdfplumber==0.10.0
# Vector database support
# pgvector==0.2.4
chromadb==1.4.1
# chromadb==0.4.22
boto3==1.42.1


transformers
torch
spacy
numpy
pandas
matplotlib
scipy

# run this once on the server: python -m spacy download en_core_web_sm